New upstream version 2.0.0~gite60d0d5

+++ winpr/libwinpr/utils/print.c
@@ -33,7 +33,7 @@
 
 #include "../log.h"
 
-void winpr_HexDump(const char* tag, int level, const BYTE* data, int length)
+void winpr_HexDump(const char* tag, UINT32 level, const BYTE* data, int length)
 {
 	const BYTE* p = data;
 	int i, line, offset = 0;
@@ -43,7 +43,7 @@ void winpr_HexDump(const char* tag, int level, const BYTE* data, int length)
 
 	if (!buffer)
 	{
-		WLog_ERR(tag, "malloc(%zd) failed with [%d] %s", blen, errno, strerror(errno));
+		WLog_ERR(tag, "malloc(%lu) failed with [%d] %s", (unsigned long)blen, errno, strerror(errno));
 		return;
 	}
 
@@ -74,7 +74,7 @@ void winpr_HexDump(const char* tag, int level, const BYTE* data, int length)
 	free(buffer);
 }
 
-void winpr_CArrayDump(const char* tag, int level, const BYTE* data, int length, int width)
+void winpr_CArrayDump(const char* tag, UINT32 level, const BYTE* data, int length, int width)
 {
 	const BYTE* p = data;
 	int i, line, offset = 0;
@@ -84,7 +84,7 @@ void winpr_CArrayDump(const char* tag, int level, const BYTE* data, int length, 
 
 	if (!buffer)
 	{
-		WLog_ERR(tag, "malloc(%zd) failed with [%d] %s", llen, errno, strerror(errno));
+		WLog_ERR(tag, "malloc(%lu) failed with [%d] %s", (unsigned long)llen, errno, strerror(errno));
 		return;
 	}
